# openapi.yaml
openapi: 3.0.3
info:
  title: 个性化新闻智能知识库 API
  version: 1.0.0
  description: 基于 Flask 的后端服务，支持知识库、检索、分析与配置。
servers:
  - url: http://localhost:5000
    description: 本地开发
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        msg:
          type: string
          example: ok
        data:
          type: object
...
    Article:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        content: { type: string }
        source_url: { type: string }
        source_name: { type: string }
        published_at: { type: string, format: date-time }
        category: { type: string }
        tags:
          type: array
          items: { type: string }
        status: { type: string, enum: [active, archived, deleted] }
        importance_score: { type: number, format: float }
...
    SearchResult:
      type: object
      properties:
        id: { type: string, description: 文档或分片ID }
        score: { type: number }
        title: { type: string }
        snippet: { type: string }
        source_url: { type: string }
        source_name: { type: string }
    QaAnswer:
      type: object
      properties:
        answer: { type: string }
        citations:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              title: { type: string }
              source_url: { type: string }
              snippet: { type: string }
paths:
...
  /api/kb/items:
    get:
      summary: 知识库列表
      parameters:
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: tag
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Article' }
                      total: { type: integer }
    post:
      summary: 批量入库（结构化数据）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/Article'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/search/semantic:
    post:
      summary: 语义搜索（相似度检索+重排）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string }
                top_k: { type: integer, default: 5 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/SearchResult' }
  /api/search/qa:
    post:
      summary: RAG 问答（含引用）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string }
                top_k: { type: integer, default: 5 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    $ref: '#/components/schemas/QaAnswer'
...